feat(summon_homun): add instruction editor and refactor into modular structure
Add multiline instruction editor using bubbles/textarea that allows users to
edit report/instructions.md before running the Docker container. If the file
doesn't exist, a helpful template is provided.

Refactored the monolithic main.go (592 lines) into clean, modular components:
- model.go: State definitions and model struct
- styles.go: Lipgloss styles and constants
- view.go: View rendering logic
- update.go: Update logic and state transitions
- validation.go: Environment and input validation
- docker.go: Docker operations
- git.go: Git repository operations
- instructions.go: Instruction file I/O
- main.go: Minimal entry point (25 lines)

Added comprehensive test coverage:
- validation_test.go: Branch name validation tests
- docker_test.go: Docker image checking tests
- git_test.go: Git operations tests
- instructions_test.go: File I/O tests

All tests passing with 100% success rate.
